updated format

+++ src/BurnToken.sol
@@ -6,24 +6,19 @@ import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {ERC20Burnable} from "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
 
 contract BurnToken is ERC20, ERC20Burnable, Ownable {
-    uint public constant BURN_RATE = 2000; // 2% burn rate
-    uint public constant BASE = 10000;
-    uint public totalBurned; // total number of tokens burned
-    mapping(address => uint) public burnedPerAddress; // the amount of tokens a user burned
+    uint256 public constant BURN_RATE = 2000; // 2% burn rate
+    uint256 public constant BASE = 10000;
+    uint256 public totalBurned; // total number of tokens burned
+    mapping(address => uint256) public burnedPerAddress; // the amount of tokens a user burned
 
-    constructor(
-        address initialOwner
-    ) ERC20("BurnToken", "BURN") Ownable(initialOwner) {
+    constructor(address initialOwner) ERC20("BurnToken", "BURN") Ownable(initialOwner) {
         _mint(initialOwner, 1000 * (10 ** decimals()));
     }
 
-    function transfer(
-        address recipient,
-        uint amount
-    ) public override returns (bool) {
-        uint rawAmount = amount * (10 ** decimals());
-        uint burnAmount = calculateBurn(rawAmount);
-        uint sendAmount = rawAmount - burnAmount;
+    function transfer(address recipient, uint256 amount) public override returns (bool) {
+        uint256 rawAmount = amount * (10 ** decimals());
+        uint256 burnAmount = calculateBurn(rawAmount);
+        uint256 sendAmount = rawAmount - burnAmount;
 
         _burn(msg.sender, burnAmount);
         burnedPerAddress[msg.sender] += burnAmount;
@@ -32,14 +27,10 @@ contract BurnToken is ERC20, ERC20Burnable, Ownable {
         return super.transfer(recipient, sendAmount);
     }
 
-    function transferFrom(
-        address sender,
-        address recipient,
-        uint amount
-    ) public override returns (bool) {
-        uint rawAmount = amount * (10 ** decimals());
-        uint burnAmount = calculateBurn(rawAmount);
-        uint sendAmount = rawAmount - burnAmount;
+    function transferFrom(address sender, address recipient, uint256 amount) public override returns (bool) {
+        uint256 rawAmount = amount * (10 ** decimals());
+        uint256 burnAmount = calculateBurn(rawAmount);
+        uint256 sendAmount = rawAmount - burnAmount;
 
         _burn(sender, burnAmount);
         burnedPerAddress[sender] += burnAmount;
@@ -48,7 +39,7 @@ contract BurnToken is ERC20, ERC20Burnable, Ownable {
         return super.transferFrom(sender, recipient, sendAmount);
     }
 
-    function calculateBurn(uint amount) internal pure returns (uint) {
+    function calculateBurn(uint256 amount) internal pure returns (uint256) {
         return (amount * BURN_RATE) / BASE;
     }
 }
